refactor: move functions to appropriate class
