+++ src/main/java/tools/Utility.java
@@ -1,4 +1,11 @@
 package tools;
 
 public class Utility {
+	public static String repeatString(int count, String str) {
+		StringBuilder result = new StringBuilder();
+		for (int cnt = 0; cnt < count; cnt++) {
+			result.append(str);
+		}
+		return result.toString();
+	}
 }
